requirements.txt
APScheduler~=3.11.0
SQLAlchemy~=2.0.40
python-dotenv~=1.1.0
pytz~=2025.2